Changes for CJ
CJ had some troubles.

1. I fixed the date to say "Match Date" and it automatically adjusts to
the previous Tuesday (but you can still override it).

2. High Run should now reset to zero for each match.

3. The Sequence # is to keep the handicaps correct when they play more
than one match per night. If you win then lose, that is slightly
different than if you lose then win. It really doesn't matter much in
the grand scheme of things.

4. I'm working on the Adjust Handicap field to automatically set the old
handicap value. Seems to work, but it is hard for me to test it.

5. You didn't lose any data when you hit Reset. It just clears the
current screen. Not sure how to repair this one.

6. I removed and added back in Steve Godwin, with a handicap of 598. And
I added in his score 80-45 over Shelby, in an 80-60 race.